+++ src/main/java/com/leander/cinema/dto/CustomerDto/movieDto/MovieResponseDto.java
@@ -4,5 +4,5 @@ public record MovieResponseDto(
         String title,
         String genre,
         int ageLimit,
-        double duration) {
+        int duration) {
 }
